Внести правки

+++ src/main/java/jm/task/core/jdbc/dao/UserDaoHibernateImpl.java
@@ -4,6 +4,7 @@ import jm.task.core.jdbc.model.User;
 import jm.task.core.jdbc.util.Util;
 import org.hibernate.Session;
 import org.hibernate.SessionFactory;
+import org.hibernate.Transaction;
 import org.hibernate.query.Query;
 
 import java.util.List;
@@ -19,53 +20,78 @@ public class UserDaoHibernateImpl implements UserDao {
 
     @Override
     public void createUsersTable() {
+        Transaction transaction = null;
         try (Session session = sessionFactory.openSession()) {
-            session.beginTransaction();
-            String request = "CREATE TABLE IF NOT EXISTS Users (" +
+            transaction = session.beginTransaction();
+            String request = "CREATE TABLE IF NOT EXISTS users (" +
                     "id BIGSERIAL PRIMARY KEY," +
                     "name VARCHAR(225) NOT NULL," +
                     "last_name VARCHAR(255) UNIQUE NOT NULL," +
                     "age SMALLINT NOT NULL" +
                     ");";
             session.createNativeQuery(request).executeUpdate();
+            session.getTransaction().commit();
+        } catch (Exception e) {
+            e.getMessage();
+            if (transaction != null)
+                transaction.rollback();
         }
     }
 
     @Override
     public void dropUsersTable() {
+        Transaction transaction = null;
         try (Session session = sessionFactory.openSession()) {
-            session.beginTransaction();
+            transaction = session.beginTransaction();
             String request = "DROP TABLE IF EXISTS users";
             session.createNativeQuery(request).executeUpdate();
+            session.getTransaction().commit();
+        } catch (Exception e) {
+            e.getMessage();
+            if (transaction != null)
+                transaction.rollback();
         }
     }
 
     @Override
     public void saveUser(String name, String lastName, byte age) {
+        Transaction transaction = null;
         try (Session session = sessionFactory.openSession()) {
-            session.beginTransaction();
+            transaction = session.beginTransaction();
             session.save(new User(name, lastName, age));
             session.getTransaction().commit();
+        } catch (Exception e) {
+            e.getMessage();
+            if (transaction != null) {
+                transaction.rollback();
+            }
         }
     }
 
     @Override
     public void removeUserById(long id) {
+        Transaction transaction = null;
         try (Session session = sessionFactory.openSession()) {
-            session.beginTransaction();
+            transaction = session.beginTransaction();
             User user = session.get(User.class, id);
-            session.delete(user);
-            session.getTransaction().commit();
+            if (user != null) {
+                session.delete(user);
+                session.getTransaction().commit();
+            } else throw new RuntimeException("Пользователь не найден.");
+        } catch (Exception e) {
+            e.getMessage();
+            if (transaction != null) {
+                transaction.rollback();
+            }
         }
     }
 
     @Override
     public List<User> getAllUsers() {
-        List<User> users;
         try (Session session = sessionFactory.openSession()) {
             session.beginTransaction();
             Query<User> query = session.createQuery("from User", User.class);
-            return users = query.list();
+            return query.list();
         }
     }
